requirements.txt
np_utils
scipy 
graphviz
keras
matplotlib
h5py
tensorflow
pillow
